+++ roseau/load_flow/exceptions.py
@@ -41,6 +41,8 @@ class RoseauLoadFlowExceptionCode(StrEnum):
     BAD_Y_SHUNT_VALUE = auto()
     BAD_TYPE_NAME_SYNTAX = auto()
     BAD_LENGTH_VALUE = auto()
+    BAD_MAX_CURRENTS = auto()
+    BAD_MAX_CURRENTS_SIZE = auto()
 
     # Transformer
     BAD_TRANSFORMER_ID = auto()