+++ scripts/update_catalogue_networks.py
@@ -101,7 +101,7 @@ if __name__ == "__main__":
                     z_line=line.parameters.z_line,
                     y_shunt=line.parameters.y_shunt,
                     # Add missing data from the IEC catalogue
-                    max_current=iec_params.max_current,
+                    max_currents=iec_params.max_currents,
                     line_type=iec_params.line_type,
                     conductor_type=iec_params.conductor_type,
                     insulator_type=iec_params.insulator_type,